Add support for preseed uploads.
This commit implements a new module for adding
support to upload a preseed file given a name/path.

For now, If the given name already exists (used as the primary key)
then the upload is ignored, if the name doesn't exists yet
then the new preseed is uploaded from the given local path.